Updated to work with new Elementor settings method.
# custom-elementor-element

This commit allows you to use the code given to immediately install and
make use of the boilerplate plugin.

Two things have been updated from the current code:

1. Added functionality to support fetching the `my-widget.php` file
from the plugin directory, instead of relying on the theme folder
having the `my-widget.php` file located in
`plugins/elementor/my-widget.php`.

2. Updates the function of getting the values of the input fields,
making use of the `$this->get_settings();` method. This is then applied
to the `$custom_text` and `$post_count` statements to evaluate
correctly.

3. Added in comments for the use of Font Awesome iconset [linking to
the Cheat Sheet](http://fontawesome.io/cheatsheet/).